Added needler
The needler is essentially a pneumatic weapon that fires small wooden or metal stakes; it holds fewer rounds than the helsing, but uses magazines that take a while to reload instead of having to bear the long reload time every time. Meant to be less accurate than the helsing but more portable and easier to fire repeatedly.